Graph any 3D function and play around with it.
This program is still a working progress and currently can only handle
one object/graph at a time. It can not properly display multiple objects
on the screen at one time.

The user interface consists of a turtle graphics display on the left
that is 600 by 600 pixels and displays the graph of the function entered
on the bottom right. The panel on the right is split into two sections.
The top section allows the user to rotate, zoom and move the three
dimensional space on the left; the bottom panel (“Object Options”)
allows the user to move the graph itself within the space (i.e. the
graph will move with respect to the grid and the grid will stay still
with respect to the screen). The functions in the “Objects Options”
panel allows the user to rotate, reflect, dilate, and stretch the graph
with respect to a specified center in addition to moving along any of
the three axes. For example, if the user were to enter a dilation factor
of 2 at a center of (300,300,300), the graph will expand out from the
upper right corner. Moving the object in the x, y, or z direction can be
done by typing the units in or using the plus and minus buttons.